implemented request validation

+++ app.py
@@ -45,23 +45,30 @@ def api_create_product():
             400,
         )
 
-    product = Product(      # instance of a product object
-        name=data["name"],
-        price=price,
-        quantity=quantity,
+    product = Product(          # instance of a product object
+        name = data["name"],
+        price = price,
+        quantity = quantity,
     )
-    db.session.add(product)  # product added to the database
+    db.session.add(product)     # product added to the database
     db.session.commit()
     return "Item added to the database"
 
 
 @app.route("/api/product/<string:name>", methods=["DELETE"]) # deletes product from db
-    # TODO add validation to check if product exists in database before deletion (return with 404)
-
 def api_delete_product(name):
-    product = db.session.get(Product, name) # gets product from database
+
+    # checks if the item exists in the products table before deletion
+    product_table = Product.query.all()                                 # dict containing all products in products table (name, price, quantity)
+    product_list = [ product.name for product in product_table]       # List product name from product table
+    
+    if name.lower() not in product_list:
+        return f"Item: {name} does not exist", 400
+
+    product = db.session.get(Product, name) 
     db.session.delete(product)
     db.session.commit()
+
     return f"{name.lower()} deleted from database"
 
 
@@ -87,9 +94,11 @@ def api_put_product(name):
         )
 
     product = db.session.get(Product, name.lower()) # gets product from database
+    
     product.price = price   # updates price
     product.quantity = quantity # updates quantity
     db.session.commit()
+
     return f"Product '{name.lower()}' has been Edited"
 
 @app.route("/api/order/<int:order_id>", methods=["GET"])
@@ -109,7 +118,6 @@ def app_get_order(order_id):
         "customer_address" : order.address,
         "completed" : order.completed,
         "products" : products_list,
-        # TODO "price" : sum(product["price"] for product in products_order)
     }
     return jsonify(res)
 
@@ -117,59 +125,78 @@ def app_get_order(order_id):
 def api_create_order():
     data = request.json # json request as a dictionary
 
-    # TODO Validate json request data
+    products = data["products"] # dict of products from post request
+    product_table = Product.query.all() # dict containing all of the products in the products table (name, price, quantity)
 
-    with app.app_context():
-        new_order = Order(name=str(data["name"]), address=str(data["address"]))
-        db.session.add(new_order)
-        db.session.commit()
-        print("New order, with ID", new_order.id)
+    # checks if all items from customer exists in product table
+    for item in products:
+        if item["name"] not in [ product.name for product in product_table]:       # creates a list of products
+            return f"Item: {item['name']} is not available or in stock", 400
 
-        # adds five random products with random quantities to the order
-        products = random.sample(Product.query.all(), k=5)
+    # checks if quantity of products is a whole number and greater than 0
+    if type(item["quantity"]) != int or item["quantity"] < 0:
+        return "Quantity must be a whole number and greater than 0"
+     
+    # creates a new order and adds order to order table
+    new_order = Order(
+        name=str(data["name"]), 
+        address=str(data["address"])
+    )
+    db.session.add(new_order)
+    db.session.commit()
+    print("New order, with ID", new_order.id)
 
-    for p in products:
-        quantity = random.randint(1, 10)
-        association = ProductsOrder(product=p, order=new_order, quantity=quantity)
+    # converts Product JSON data to a list of instances
+    products_ordered = [Product(name=str(item["name"]), price=float(0.00), quantity=int(item["quantity"])) 
+                        for item in data["products"]]
+
+    # adds products selected by customer into database associated with customer order
+    for item in products_ordered:
+        association = ProductsOrder(product_name=item.name, order=new_order, quantity=item.quantity)
         db.session.add(association)
-
     db.session.commit()
 
-    return "Order added to the database"
 
-@app.route("/api/product/<int:order_id>", methods=["PUT"])
+    return f"Order added to the database {data}"
+
+
+
+@app.route("/api/order/<int:order_id>", methods=["PUT"])
 def api_process_order(order_id):
     customer_order = db.session.get(Order, int(order_id))
-    data = request.json # json request as a dictionary
+
+    data = request.json # json request as a dictionary    
     order_status = data["process"]
 
+    # checks for any errors
+    if request.json is None:
+        return "The JSON provided is invalid (missing)", 400
+
     if order_status.upper() != "TRUE":
-        return f"The JSON provided is invalid (missing: {order_status})", 400
+        return f"The JSON provided is invalid (invalid: {order_status})", 400
     
-    # checks if order has already been processed
-    if customer_order.completed == bool("TRUE"):
+    if customer_order.completed:
         return f"Customer order ID: {order_id} has already been processed"    
 
-    
-    products_order = ProductsOrder.query.all() # dict of all queries of the order in the db
-    products_inventory = Product.query.all() # dict all queries of the order in the db
 
-    order = [{ "name" : product.product_name, "quantity" : product.quantity } # Dict of products from order
+    products_order = ProductsOrder.query.all()          # dict of all queries of the order in the db
+    products_inventory = Product.query.all()          # dict all queries of the order in the db
+
+    order = [{ "name" : product.product_name, "order_id" : product.order_id, "quantity" : product.quantity } # Dict of products from order
                      for product in products_order]
     
     inventory = [ {"name" : product.name, "quantity" : product.quantity} # Dict of products from inventory
                       for product in products_inventory]
-    
-    
-    # if the order contains more products than available, adjust the order accordingly. (manipulates order & inventory dicts)
+
+    # if the order contains more products than available, adjusts the order accordingly. 
+    # (manipulates order & inventory dicts)
     for item in order:
         product = next((p for p in inventory if p["name"] == item["name"]), None)
-        if product:
+        if product and item["order_id"] == order_id:  
             item_quantity = min(item["quantity"], product["quantity"])
             item["quantity"] = item_quantity
             product["quantity"] -= int(item_quantity)
 
-    
     for item in inventory:
         for product in order:
                 if product["name"] == item["name"]: 
@@ -184,6 +211,7 @@ def api_process_order(order_id):
     db.session.commit()
 
     return f"Order ID: {order_id} has been processed"
+
 
 
 if __name__ == "__main__":
